backend/requirements.txt
Flask==2.3.3
Flask-CORS==4.0.0
Flask-SQLAlchemy==3.0.5
python-dotenv==1.0.0
Werkzeug==2.3.7
SQLAlchemy==2.0.21
gunicorn==21.2.0
pytest==7.4.2
requests==2.31.0
